refactor: Improve user search functionality in account and user sections
This commit updates the code in the account and user sections to improve the user search functionality. In the account-list-view.tsx file, the applyFilter function is modified to include a search by email in addition to the search by username. This change allows users to search for accounts using either the username or email. In the user-table-filters-result.tsx file, a commented out line of code is removed, which was not being used and was unnecessary. These changes enhance the user search experience and provide more flexibility in finding specific users.