refactor: Add deleteMutate prop to PostListHorizontal and PostItemHorizontal components
This commit adds the deleteMutate prop to the PostListHorizontal and PostItemHorizontal components. The deleteMutate prop is a function that allows for deleting a post from the list. This change enables the delete functionality in the blog section, improving the user experience by providing the ability to delete posts directly from the list view.